% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_bad_experiment_msg.R
\name{cm_bad_experiment_msg}
\alias{cm_bad_experiment_msg}
\title{Compose an error message}
\usage{
cm_bad_experiment_msg(thisExperiment)
}
\arguments{
\item{thisExperiment}{A cm_experiment object}
}
\value{
A character string representing a composite error message detailing all errors found by examining the status field of the cm_experiment object
}
\description{
Compose an error message
}
