% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_create_new_experiment.R
\name{cm_create_new_experiment}
\alias{cm_create_new_experiment}
\title{Create a new cm experiment}
\usage{
cm_create_new_experiment(userID = "", experimentName = "")
}
\arguments{
\item{userID}{Character. User ID}

\item{experimentName}{Character. Name of the experiment assigned by the user}
}
\value{
An initialised cm_experiment object
}
\description{
Create a new cm experiment
}
\examples{
\dontrun{
thisExperiment <- cm_create_new_experiment("user123", "Species assemblages model 1")
}
}
