% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_gdm_pcaPlot.R
\name{cm_gdm_pcaPlot}
\alias{cm_gdm_pcaPlot}
\title{GDM PCA plot of scaled covariates}
\usage{
cm_gdm_pcaPlot(thisExperiment, outFolder = "", trace = FALSE)
}
\arguments{
\item{thisExperiment}{Object of class 'cm_experiment'}

\item{outFolder}{Character (string). Path of a folder into which the output GIS layer and PNG image will be written}

\item{trace}{Logical. Produce console messages? Default is FALSE}
}
\value{
Nothing
}
\description{
Produce a raster map colour-coded using the first three components of a Principal Component Analysis (PCA) of scaled covariates. Grid cells with the similar colours represent cells with the same combination of scaled environmental variables and therefore similar predicted community composition.
}
\details{
Both graphics output (as a PNG-formatted image) and a GeoTIFF raster are output.
}
\examples{
\dontrun{
cm_gdm_pcaPlot(myExperiemnt, outFolder = "exprimentResults")
}
}
