% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_gdm_report.R
\name{cm_gdm_report}
\alias{cm_gdm_report}
\title{GDM report}
\usage{
cm_gdm_report(thisExperiment, outFile = "", format = "pdf")
}
\arguments{
\item{thisExperiment}{cmGDM object}

\item{outFile}{Character. Name of the file (including a fully specified path) to which outpout will be written.}

\item{format}{Character. Report format to be used. Default is "pdf". Other optioins are "word" and "html".}
}
\value{
Nothing
}
\description{
GDM report
}
\examples{
\dontrun{

cm_gdm_report(myExperiment)

}
}
