% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_gdm_summary.R
\name{cm_gdm_summary}
\alias{cm_gdm_summary}
\title{GDM summary}
\usage{
cm_gdm_summary(thisExperiment, outFile = "")
}
\arguments{
\item{thisExperiment}{cm_experiment object. Experiment object storing a successfully fitted GDM}

\item{outFile}{Character. File name (with full path) into which the summary will be written. Default of "" means no file will be written}
}
\value{

}
\description{
Generate a summary report on a fitted GDM. Output is always written to the console and, optionally, also written to a text file.
}
\examples{
\dontrun{
# Console-only output
cm_gdm_summary(anExperiment)

# Console and file output
cm_gdm_summary(anotherExperiment, file = "thisFile.txt")
}

}
