% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_load_pred_covars.R
\name{cm_load_prediction_data}
\alias{cm_load_prediction_data}
\title{Load Prediction Covariate Data}
\usage{
cm_load_prediction_data(
  thisExperiment,
  src_folder = "",
  pred_filenames = "",
  label = "",
  trace = FALSE
)
}
\arguments{
\item{thisExperiment}{cm_experiment object. Object for the current experiment. NOTE: Require site data to have been successfully added to the experiment before this method is called}

\item{src_folder}{Character. Path to the folder storing raster layers}

\item{pred_filenames}{Character vector. Vector giving the names of the files in src-folder to be used as prediction covariates in this experiment}

\item{label}{Character. An optional human-friendly label for the project's environmental data}

\item{trace}{Logical. Produce helpful diagnostic messages? Default is FALSE, therefore radio silence is maintained until told otherwise}
}
\value{

}
\description{
Load metadata for GIS layers to be used to predict dissimilarities between sites/samples under different environmental conditions. Checks are made to ensure that the prediction data matches the covariates used to fit a GDM
}
\examples{
\dontrun{
# Assume a previously created cm_experiment object 'Butterfly_effect'

dataFolder <- "/path/to/GIS_layers"

# Make character vector of ALL GIS layer files in the source folder
#### NOTE: File names MUST be the same as loaded by cm_load_covars
#### even though they may cover different regions and/or times
theFiles <- list.files(dataFolder, '*.*')
Butterfly_effect <- cm_load_pred_data(Butterfly_effect,
                                      dataFolder,
                                      theFiles)

# Load only selected subset of files:
theFiles <- c("bio01.tif", "bio05.tif", "TPI.tif")
Butterfly_effect <- cm_load_pred_data(Butterfly_effect,
                                      dataFolder,
                                      theFiles)
}
}
