% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_load_site_table.R
\name{cm_load_site_table}
\alias{cm_load_site_table}
\title{Load a site table and perform checks}
\usage{
cm_load_site_table(
  thisExperiment = NULL,
  siteFilename = NULL,
  sheet = NULL,
  siteCol = NULL,
  longitudeCol = NULL,
  latitudeCol = NULL,
  weightType = "equal",
  weightsCol = NULL,
  outFolder = "~/cmGDM/",
  trace = FALSE
)
}
\arguments{
\item{thisExperiment}{cm_experiment. Object to which the site data will be added when checks are passed}

\item{siteFilename}{Character. Full path to the site file to be added to the cm_experiment object}

\item{sheet}{Character or integer. Name or numeric index to the sheet holding the site information when the file is a spreadsheet/workbook}

\item{siteCol}{Character or integer. Name or numeric index to the column holding the site names or identifiers}

\item{longitudeCol}{Character or integer. Name or numeric index to the column holding longitude values}

\item{latitudeCol}{Character or integer. Name or numeric index to the column holding latitude values}

\item{weightType}{Character. Weight method to be applied to sites. Default = "equal", "richness" adds weights equal to the species richness at each site, and "custom" uses site weights supplied in the data table in the column given parameter \emph{weightsCol}.}

\item{weightsCol}{Character or integer. Name of numeric index (column number) of the column supplying custom weights for each site}

\item{outFolder}{String. Path of the output folder}

\item{trace}{Logical. Produce helpful diagnostic messages? Default is FALSE, therefore radio silence is maintained until told otherwise}
}
\value{
cm_experiment object with updated fields and stores a copy in the user's project folder for this experiment
}
\description{
Load a site table and perform checks
}
\examples{
\dontrun{
# Create a cm_experiment object:
birdTurnover1 <- cm_create_new_experiment("user456", "Bird turnover GDM")

# CSV-delimited text file:
birdTurnover1 <- cm_load_site_table(birdTurnover1,
                                    "/path/to/siteData.csv",
                                    siteCol = 1,
                                    longitudeCol = 2,
                                    latitudeCol = 3)
                                    
# Excel spreadsheet (sheet must be given as sheet number or sheet name):
birdTurnover1 <- cm_load_site_table(birdTurnover1,
                                    "/path/to/siteData.xlsx",
                                    sheet = "Sites",
                                    siteCol = 1,
                                    longitudeCol = 2,
                                    latitudeCol = 3)
}
}
