% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_gdm_prediction.R
\name{cm_predict_gdm}
\alias{cm_predict_gdm}
\title{Make predictions of change in composition}
\usage{
cm_predict_gdm(thisExperiment, outFolder = "~/cmGDM/", trace = FALSE)
}
\arguments{
\item{thisExperiment}{cm_experiment. Experiment object to be used in making predictions}

\item{outFolder}{String. Path of the output folder}

\item{trace}{Logical. Should informative progress messages be output}
}
\value{
Returns an updated cm_experiment object and saves it to the experiment folder
}
\description{
Make predictions of change in composition
}
\examples{
\dontrun{
  cm_predict_gdm(myExperiment, outFolder = "path/to/output")
}
}
