% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_run_gdm_experiment.R
\name{cm_run_gdm_experiment}
\alias{cm_run_gdm_experiment}
\title{Run Community Modelling GDM experiment}
\usage{
cm_run_gdm_experiment(thisExperiment, trace = FALSE)
}
\arguments{
\item{thisExperiment}{Character. Name of this experiment}

\item{trace}{Logical. Produce helpful diagnostic messages? Default is FALSE, therefore radio silence is maintained until told otherwise}
}
\value{
cm_experiment object with updated fields
}
\description{
Experiments can only be run when all preceding steps in the workflow have been successfully completed.
}
\examples{
\dontrun{

anotherExperiment <- cm_run_gdm_experiment(anotherExperiment)

}
}
