% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_show_extrapolation.R
\name{cm_show_extrapolation}
\alias{cm_show_extrapolation}
\title{Show extrapolation}
\usage{
cm_show_extrapolation(
  thisExperiment,
  covarSrc = "train",
  showVarImp = c("all", "important"),
  allPlots = FALSE,
  outFolder = "~/cmGDM/"
)
}
\arguments{
\item{thisExperiment}{cm_experiment object. The experiment with a fitted GDM to be processed}

\item{covarSrc}{Character string. Which covariate data set should be used to test for extrapolation? Default is "train', i.e. the data set used to fit or 'train' the GDM. Alternative is 'predict', which uses the prediction data set after first check that it has been uploaded by the user}

\item{showVarImp}{Character string. Show extrapolation for all covariates presented to model fitting ("all"), or just those with variable importance > 0 ("important")}

\item{allPlots}{Logical. Show extrapolation plots for each covariate plus the overall extrapolation plot (TRUE) or just the overall plot (default, FALSE)}

\item{outFolder}{Character string. Full path to a folder to be used for output geoTIFF and PNG files}
}
\value{
Nothing
}
\description{
Models applied to covariate values outside the range used to fit the model cannot be trusted.
}
\details{
EXTREME caution is required when interpreting results in areas where extrapolation has occurred.

This function identifies the regions where extrapolation has occurred, and outputs PNG and geoTIFF files showing grid cells within which extrapolation has NOT occurred (i.e. "good" cells)
}
\examples{
\dontrun{
# Show areas with NO extrapolation for all variables
cm_show_extrapolation(myExperiment, showVarImp = "all")

# Show areas with No extrapolation but only for IMPORTANT variables
cm_show_extrapolation(myExperiment, showVarImp = "important")

# Show extrapolation for important varaible using hte prediction data set
cm_show_extrapolation(myExperiment, covarSrc = "predict", showVarImp = "important")
}
}
