% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_status_check.R
\name{cm_status_check}
\alias{cm_status_check}
\title{EcoCommons Community Model Object Status Summary}
\usage{
cm_status_check(thisExperiment)
}
\arguments{
\item{thisExperiment}{cm_experiment. The experiment object whose status is to be summarised}
}
\value{
Nothing
}
\description{
A short dump of status information for a cm_experiment object
}
