% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmGDM.R
\docType{package}
\name{cmGDM}
\alias{cmGDM}
\title{cmGDM: Community Modelling using Generalised Dissimilarity Modelling}
\description{
EcoCommons Community Modelling experiments run using Generalised Dissimilarity Modelling
}
