% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_bad_experiment_msg.R
\name{cm_bad_experiment_msg}
\alias{cm_bad_experiment_msg}
\title{Compose an error message}
\usage{
cm_bad_experiment_msg(thisExperiment)
}
\arguments{
\item{thisExperiment}{A cm_experiment object created by a call to \link{cm_create_new_experiment} and updated by each step in the workflow}
}
\value{
A character string representing a composite error message detailing all errors found by examining the \emph{status} field of the cm_experiment object
}
\description{
Prepare a multi-part message informing the user of reasons for failure to allow a requested action such as fitting a GDM using a cm_experiment object
}
