% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_gdm_report.R
\name{cm_gdm_report}
\alias{cm_gdm_report}
\title{GDM report}
\usage{
cm_gdm_report(
  thisExperiment,
  outFolder = "~/cmGDM/",
  format = "pdf",
  trace = FALSE
)
}
\arguments{
\item{thisExperiment}{cm_experiment. A cm_experiment object to be used to generate the report}

\item{outFolder}{String. Base path to which output will be written.}

\item{format}{String. Report format to be used. Default is "pdf" which is the only formatted implement at present. Other options will include "word" and "html".}

\item{trace}{Logical. Produce hopefully diagnostic messages. Default is to remain silent.}
}
\value{
Nothing but with a side-effect of saving the report pdf to the project folder
}
\description{
A pdf file is produced documenting the outcomes of the model fitted using the cm_experiment passed in \emph{thisExperiment}.
}
\details{
Note that all graphical elements included in the report are also separately stored in the user's project folder for the experiment, and are therefore \emph{potentially} downloadable so that the user may incorporate them in other documents such as reports or theses.
}
\examples{
\dontrun{

cm_gdm_report(myExperiment)

}
}
