% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_performance_plots.R
\name{cm_performance_plots}
\alias{cm_performance_plots}
\title{GDM performance plots}
\usage{
cm_performance_plots(
  thisExperiment,
  outFolder = "~/cmGDM/",
  showVarImp = "all"
)
}
\arguments{
\item{thisExperiment}{cm_experiment object. An experiment object storing a successfully fitted GDM}

\item{outFolder}{Character. Path to user's experiment folder}

\item{showVarImp}{Character. Control which set of variable contribution plots are created: "all" generates all plots; "nonzero" generates only plots for variables with a variable importance score greater than zero. If variable importance permutation has not been run, this parameter will be ignored}
}
\value{
Nothing
}
\description{
Create three types of performance evaluation plot for a fitted GDM: observed v predicted dissimilarity, ecological distance v observed dissimilarity, and variable contribution plots
}
\examples{
\dontrun{
# Default plots: show ALL variable contribution plots
cm_performance_plots(anExperiment,
                     "/path/to/my/stuff")
        
# Show only var. contrib.plots when varaible importance is > 0
cm_performance_plots(anExperiment,
                     "/path/to/my/stuff",
                     "nonzero")
}
}
