% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_show_extrapolation.R
\name{cm_show_extrapolation}
\alias{cm_show_extrapolation}
\title{Show extrapolation}
\usage{
cm_show_extrapolation(
  thisExperiment,
  extrapType = c("all", "important"),
  allPlots = FALSE,
  outFolder = ""
)
}
\arguments{
\item{thisExperiment}{cm_experiment object. The experiment with a fitted GDM to be processed}

\item{extrapType}{Character string. Show extrapolation for all covariates presented to model fitting ("all"), or just those with variable importance > 0 ("important")}

\item{allPlots}{Logical. Show extrapolation plots for each covariate plus the overall extrapolation plot (TRUE) or just the overall plot (default, FALSE)}

\item{outFolder}{Character string. Full path to a folder to be used for output geoTIFF and PNG files}
}
\value{

}
\description{
Models applied to covariate values outside the range used to fit the model cannot be trusted. EXTRREME caution is required when interpreting results in areas where extrapolation has occurred. This function identifies the regions where extrapolation has occurred, and outputs PNG and geoTIFF files showing grid cells within which extrapolation has NOT occurred (i.e. "good" cells)
}
